+++ CHANGELOG.md
@@ -11,7 +11,11 @@ All user visible changes to this project will be documented in this file. This p
 
 [Diff](/../../compare/3aa35d5bf8ba9728f54db7bf4e21425711097cda...v0.1.0-alpha.6)
 
+### Added
 
+- UI:
+    - Chat page:
+        - Message forwarding (!138, #104).
 
 
 [Semantic Versioning 2.0.0]: https://semver.org